changing searchUserForm to a more generic searchForm
