+++ webapp/src/pages/index.ts
@@ -1,5 +1,5 @@
 export { default as Landing } from './Landing/Landing';
-export { default as Users } from './Users/Users';
+export { Users } from './Users/Users';
 export { default as Play } from './Play/Play';
 export { default as Chat } from './Chat/Chat';
 export { default as ComponentsBook } from './ComponentsBook/ComponentsBook';
