+++ webapp/src/shared/components/Input/SearchForm.css
@@ -1,4 +1,4 @@
-.search-user-form {
+.search-form {
   display: flex;
   align-items: center;
 }
