+++ webapp/src/shared/components/index.ts
@@ -27,4 +27,4 @@ export { default as RowsList } from './RowsList/RowsList';
 export * from './RowsList/RowsList';
 
 export { default as EditUserForm } from './Input/EditUserForm';
-export { default as SearchUserForm } from './Input/SearchUserForm';
+export { default as SearchUserForm } from './Input/SearchForm';
